pyproject.toml
[tool.poetry]
name = "menu"
version = "0.1.0"
description = ""
authors = ["iimgera <iimgera28062004@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django-jazzmin = "^2.6.0"
Django = "^4.2"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
drf-yasg = "^1.21.5"
django-filter = "^23.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
